chore: 내 취미 가져오는 api 위치 변경
